Update the Typeform data load in
The logic around how the typeform data was being pulled in
was incorrect so we were losing some of the words. This
should work better.

For future, we have to find the minimum date and then set
the beforeToken from there because the Typeform API only
allows 1000 responses at a time.